+++ src/main/java/lotto/domain/lotto/LottosResult.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 public class LottosResult {
     private final int[] winningStaticsArray;
