+++ src/main/java/lotto/domain/lotto/strategy/LottoGeneratingStrategy.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.lotto.strategy;
+
+import lotto.domain.lotto.Lotto;
 
 @FunctionalInterface
 public interface LottoGeneratingStrategy {
