+++ src/test/java/lotto/domain/lotto/LottosTest.java
@@ -1,5 +1,8 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
+import lotto.domain.PurchaseAmountOfMoney;
+import lotto.domain.WinningNumbers;
+import lotto.domain.lotto.strategy.LottoGeneratingStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -25,7 +28,7 @@ class LottosTest {
     @ParameterizedTest
     @CsvSource(value = {"1, 2, 3, 7, 8, 9:3:5", "1, 2, 3, 4, 8, 9:4:50", "1, 2, 3, 4, 5, 9:5:1500", "1, 2, 3, 4, 5, 6:6:2000000"}, delimiter = ':')
     @DisplayName("result(): 로또의 결과(당첨 통계, 수익률)를 반환합니다.")
-    void testResult(String winningNumberInput, int matchCount, double expectedRateOfReturn ) {
+    void testResult(String winningNumberInput, int matchCount, double expectedRateOfReturn) {
         PurchaseAmountOfMoney purchaseAmountOfMoney = PurchaseAmountOfMoney.valueOf("1000");
         Lottos lottos = Lottos.valueOf(purchaseAmountOfMoney);
         lottos.purchaseLotto(lottoGeneratingStrategyStub);
